moved and renamed lambdas
